+++ R/tableby.R
@@ -246,7 +246,8 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, strata,
     xTerms <- Map(modeldf, names(modeldf), f = function(col, nm) {
       if(is.null(nameEff <- attr(col, "name"))) nameEff <- nm
       if(is.null(labelEff <- attr(col, "label"))) labelEff <- nameEff
-      list(variable=nameEff, label=labelEff, term=nm)
+      if(is.null(termEff <- attr(col, "term"))) termEff <- nm
+      list(variable=nameEff, label=labelEff, term=termEff, term.orig = nm)
     })
     names(xTerms) <- vapply(xTerms, "[[", NA_character_, "variable")